seperation of concern and dir relocation for better readability
