--- todo_mern_client/src/components/Modal/types/inputtypes.ts
@@ -1,7 +1,0 @@
-export type Inputs = {
-  title: string | undefined;
-  description?: string;
-  dueDate?: string;
-  categoryId?: string;
-  priority?: string;
-};
