+++ todo_mern_client/src/components/index.ts
@@ -1,6 +1,5 @@
 export { default as TodoCard } from "./TodoCard/TodoCard";
 export { default as NewTaskBtn } from "./NewTaskBtn/NewTaskBtn";
-export { default as TaskCreateModal } from "./TaskCreateModal/TaskCreateModal";
 export { default as CategoriesCard } from "./CategoriesCard/CategoriesCard";
 export { default as Header } from "./Header/Header";
 export { default as Nav } from "./Nav/Nav";
